Support creating .pkg installers using the debug build configuration.
Users reporting bugs will be able to use these packages to install debug
builds of Background Music without having to install from source. This
is mainly useful because debug builds have more detailed logging.
Hopefully we'll get around to adding an option to enable debug logging
at runtime, but this should work well enough for now.

Also:
 - Use newer macOS images in Travis CI builds.
 - Fix an xcrun command in build_and_install.sh that was accidentally
   being started in the background.
 - Fix build_and_install.sh building libPublicUtility.a twice for no
   reason.

+++ package.sh
@@ -19,10 +19,11 @@
 #
 # package.sh
 #
-# Copyright © 2017 Kyle Neideck
+# Copyright © 2017, 2018 Kyle Neideck
 # Copyright © 2016, 2017 Takayama Fumihiko
 #
 # Build Background Music and package it into a .pkg file and a .zip of the debug symbols (dSYM).
+# Call this script with -d to use the debug build configuration.
 #
 # Based on https://github.com/tekezo/Karabiner-Elements/blob/master/make-package.sh
 #
@@ -46,26 +47,63 @@ set_permissions() {
 
 # --------------------------------------------------
 
-# Build
-bash build_and_install.sh -b
+# Use the release configuration by default.
+debug_build=NO
+build_output_path="build/Release"
 
+# Handle the options passed to this script.
+while getopts ":d" opt; do
+    case $opt in
+        d)
+            debug_build=YES
+            build_output_path="build/Debug"
+            ;;
+        \?)
+            echo "Invalid option: -$OPTARG" >&2
+            exit 1
+            ;;
+    esac
+done
+
+# Build
+if [[ $debug_build == YES ]]; then
+    # Disable AddressSanitizer so we can distribute debug packages to users reporting bugs without
+    # worrying about loading the AddressSanitizer dylib in coreaudiod.
+    #
+    # TODO: Would debug packages be more useful if they were built with optimization (i.e. using the
+    #       DebugOpt configuration instead of Debug)?
+    ENABLE_ASAN=NO bash build_and_install.sh -b -d
+    build_status=$?
+else
+    bash build_and_install.sh -b
+    build_status=$?
+fi
+
+# Exit if the build failed.
+if [[ $build_status -ne 0 ]]; then
+    exit $build_status
+fi
+
+# Read the version string from the build.
 version="$(/usr/libexec/PlistBuddy \
     -c "Print CFBundleShortVersionString" \
-    "BGMApp/build/Release/Background Music.app/Contents/Info.plist")"
+    "BGMApp/${build_output_path}/Background Music.app/Contents/Info.plist")"
 
 # Everything in out_dir at the end of this script will be released in the Travis CI builds.
 out_dir="Background-Music-$version"
 rm -rf "$out_dir"
 mkdir "$out_dir"
 
-# Separate the debug symbols and the .app
-echo "Archiving debug symbols"
+if [[ $debug_build == NO ]]; then
+    # Separate the debug symbols and the .app bundle.
+    echo "Archiving debug symbols"
 
-dsym_archive="$out_dir/Background Music.dSYM-$version.zip"
-mv "BGMApp/build/Release/Background Music.app/Contents/MacOS/Background Music.dSYM" \
-   "Background Music.dSYM"
-zip -r "$dsym_archive" "Background Music.dSYM"
-rm -r "Background Music.dSYM"
+    dsym_archive="$out_dir/Background Music.dSYM-$version.zip"
+    mv "BGMApp/${build_output_path}/Background Music.app/Contents/MacOS/Background Music.dSYM" \
+       "Background Music.dSYM"
+    zip -r "$dsym_archive" "Background Music.dSYM"
+    rm -r "Background Music.dSYM"
+fi
 
 # --------------------------------------------------
 
@@ -75,10 +113,11 @@ rm -rf "pkgroot"
 mkdir -p "pkgroot"
 
 mkdir -p "pkgroot/Library/Audio/Plug-Ins/HAL"
-cp -R "BGMDriver/build/Release/Background Music Device.driver" "pkgroot/Library/Audio/Plug-Ins/HAL/"
+cp -R "BGMDriver/${build_output_path}/Background Music Device.driver" \
+      "pkgroot/Library/Audio/Plug-Ins/HAL/"
 
 mkdir -p "pkgroot/Applications"
-cp -R "BGMApp/build/Release/Background Music.app" "pkgroot/Applications"
+cp -R "BGMApp/${build_output_path}/Background Music.app" "pkgroot/Applications"
 
 scripts_dir="$(mktemp -d)"
 cp "pkg/preinstall" "$scripts_dir"
@@ -86,7 +125,7 @@ cp "pkg/postinstall" "$scripts_dir"
 cp "BGMApp/BGMXPCHelper/com.bearisdriving.BGM.XPCHelper.plist.template" "$scripts_dir"
 cp "BGMApp/BGMXPCHelper/safe_install_dir.sh" "$scripts_dir"
 cp "BGMApp/BGMXPCHelper/post_install.sh" "$scripts_dir"
-cp -R "BGMApp/build/Release/BGMXPCHelper.xpc" "$scripts_dir"
+cp -R "BGMApp/${build_output_path}/BGMXPCHelper.xpc" "$scripts_dir"
 
 set_permissions "pkgroot"
 chmod 755 "pkgroot/Applications/Background Music.app/Contents/MacOS/Background Music"
@@ -133,9 +172,16 @@ rm -rf "pkgres"
 rm -f "pkg/Distribution.xml"
 
 # Print checksums
-echo "MD5 checksums:"
-md5 {"$pkg","$dsym_archive"}
-echo "SHA256 checksums:"
-shasum -a 256 {"$pkg","$dsym_archive"}
+if [[ $debug_build == YES ]]; then
+    echo "MD5 checksum:"
+    md5 "$pkg"
+    echo "SHA256 checksum:"
+    shasum -a 256 "$pkg"
+else
+    echo "MD5 checksums:"
+    md5 {"$pkg","$dsym_archive"}
+    echo "SHA256 checksums:"
+    shasum -a 256 {"$pkg","$dsym_archive"}
+fi
